backend/requirements.txt
fastapi
uvicorn
groq
python-multipart
pypdf
python-pptx
python-dotenv
requests
